requirements.txt
Django==6.0.1

django-ipware==7.0.1

Playwright== 1.57.0

requests==2.32.5

pyTelegramBotAPI==4.30.0

celery==5.6.2